Move subs to dialog
